geriatric_admin/requirements/base.txt
# Base requirements for Geriatric Administration System

# Django and core dependencies
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
django-filter>=23.0.0
django-crispy-forms>=2.0.0
crispy-bootstrap5>=0.7.0

# Database
psycopg2-binary>=2.9.0
django-environ>=0.10.0

# Cache and sessions
django-redis>=5.2.0
redis>=4.5.0

# Database connection pooling
django-db-connection-pool>=1.2.0

# Security and authentication
django-otp>=1.1.0
qrcode>=7.4.0

# File handling and storage
Pillow>=10.0.0

# API documentation
drf-spectacular>=0.26.0

# Utilities
python-decouple>=3.8.0
python-dateutil>=2.8.0
pytz>=2023.3

# Logging
python-json-logger>=2.0.0

# Encryption
cryptography>=41.0.0
django-encrypted-model-fields>=0.6.0

# Data validation
marshmallow>=3.19.0

# Excel/CSV export
openpyxl>=3.1.0
pandas>=2.0.0

# PDF generation
reportlab>=4.0.0
weasyprint>=59.0.0

# Background tasks
celery>=5.3.0
django-celery-beat>=2.5.0
django-celery-results>=2.5.0

# Monitoring and health checks
django-health-check>=3.17.0

# Environment variables
python-dotenv>=1.0.0